Seal image_ops.py's interface. Change: 138937459
